Build bundled source release
This adds a release tarball with a source code version of Akku. It
works the same, but much slower.

Also fixes the semver range used by "akku add". The ^ range picks a
version that will stay compatible by semver semantics. The ~ range is
stricter and doesn't permit newer minor versions.