Partially support symlinking files from the current project
While developing it is much more convenient if .akku/lib carries
symlinks to the current project's libraries, rather than copies.
A requirement is that files contain a single library form.

This commit contains support for doing the symlinks themselves, but
the repo scanner has to be taught about these symlinks before the
feature can be enabled.